requirements.txt
opencv_python==4.5.2.54
numpy==1.19.5
tensorflow==2.5.0
Pillow==8.3.1
